+++ packages/talker/lib/src/models/models.dart
@@ -1,1 +1,5 @@
-export 'talker_data/talker_data.dart';
+export 'talker_data.dart';
+export 'talker_data_interface.dart';
+export 'talker_error.dart';
+export 'talker_exception.dart';
+export 'talker_log.dart';