+++ packages/talker/lib/src/utils/error_handler/talker_error_handler_interface.dart
@@ -1,9 +1,0 @@
-import 'package:talker/talker.dart';
-
-abstract class TalkerErrorHandlerInterface {
-  TalkerDataInterface handle(
-    Object exception, [
-    StackTrace? stackTrace,
-    String? msg,
-  ]);
-}