+++ packages/talker/test/talker_settings_test.dart
@@ -26,27 +26,6 @@ void main() {
       );
     });
 
-    test('Equality', () async {
-      final settings = TalkerSettings(
-        useConsoleLogs: false,
-      );
-
-      final settings2 = TalkerSettings(
-        useConsoleLogs: false,
-      );
-
-      expect(settings, settings2);
-    });
-
-    test('hashCode', () async {
-      final settings = TalkerSettings(
-        useConsoleLogs: false,
-      );
-
-      expect(settings.hashCode, isNotNull);
-      expect(settings.hashCode, isNot(0));
-    });
-
     test('copyWith', () async {
       final settings = TalkerSettings();
       final newSettings = settings.copyWith(