+++ packages/talker_bloc_logger/lib/talker_bloc_logger.dart
@@ -1,5 +1,5 @@
 library talker_bloc_logger;
 
-export 'src/bloc_logs.dart';
-export 'src/talker_bloc_logger_observer.dart';
-export 'src/talker_bloc_logger_settings.dart';
+export 'bloc_logs.dart';
+export 'talker_bloc_logger_observer.dart';
+export 'talker_bloc_logger_settings.dart';