+++ packages/talker_flutter/lib/src/models/abstract/talker_flutter_adapter_interface.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// The interface forces to use
-/// the [Color] color in heir implementer class
-abstract class TalkerFlutterAdapterInterface {
-  Color? get color;
-}